fix: improve reconciliation logic to log transaction removal when balances are in sync

+++ src/utils.ts
@@ -91,23 +91,21 @@ export async function scrapeAndImportTransactions({companyId, bank}: ScrapeTrans
 
 		const existingReconciliation = allAccountTxns.find(txn => txn.imported_id === reconciliationImportedId);
 
-		// If balances are already in sync, remove any existing reconciliation and exit.
-		if (balanceDiff === 0) {
-			if (existingReconciliation) {
-				stdout.mute();
-				await actual.deleteTransaction(existingReconciliation.id);
-				stdout.unmute();
-				log('RECONCILIATION_REMOVED');
-			}
-
-			return;
-		}
-
 		log('RECONCILIATION', {
 			from: currentBalance,
 			to: accountBalance,
 			diff: balanceDiff,
 		});
+
+		// If balances are already in sync, remove any existing reconciliation and exit.
+		if (existingReconciliation && balanceDiff === 0) {
+			stdout.mute();
+			await actual.deleteTransaction(existingReconciliation.id);
+			stdout.unmute();
+			log(`RECONCILIATION_REMOVED, since ${currentBalance.toLocaleString()} === ${accountBalance.toLocaleString()}`, {transactionId: existingReconciliation.id});
+
+			return;
+		}
 
 		const reconciliationTxn = {
 			account: bank.actualAccountId,
